requirements.txt
PyMuPDF
Pillow
torchvision
torch